drops: create one base Drop class which can be set as mutable or not
